requirements.txt
asgiref==3.2.7
autopep8==1.5.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
Django==3.0.7
django-appconf==1.0.4
django-compressor==2.4
django-sass-processor==0.8
djangorestframework==3.11.0
httplib2==0.18.0
idna==2.9
libsass==0.20.0
MouseInfo==0.1.3
oauth2==1.9.0.post1
oauthlib==3.1.0
Pillow==7.1.1
PyAutoGUI==0.9.50
pycodestyle==2.6.0
pycparser==2.20
PyGetWindow==0.0.8
PyJWT==1.7.1
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-openid==2.2.5
python3-openid==3.1.0
PyTweening==1.0.3
pytz==2019.3
rcssmin==1.0.6
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.8
